Realized I should move some of the puts to just a dry run feature.

+++ src/handlers/append.cr
@@ -1,5 +1,5 @@
 class AppendHandler
-  def initialize(argument, file)
+  def initialize(argument, file, @dry_run = false)
     @appendInfo = argument.as(Array(TOML::Type))
     @cf = Common.new(file)
   end
@@ -7,8 +7,12 @@ class AppendHandler
   def append
     lineToAdd = @appendInfo[0].as(String)
     pathToFile = @appendInfo[1].as(String)
-    @cf.idatLog("Appending Line: #{lineToAdd} to: #{pathToFile}")
-    puts "Appending Line... #{lineToAdd.colorize.mode(:bold)} to #{pathToFile.colorize.mode(:bold)}"
-    File.write(pathToFile, lineToAdd, mode: "a")
+    if @dry_run
+      puts "[dry-run] Would append: #{lineToAdd.colorize.mode(:bold)} to #{pathToFile.colorize.mode(:bold)}"
+    else
+      @cf.idatLog("Appending Line: #{lineToAdd} to: #{pathToFile}")
+      puts "Appending Line... #{lineToAdd.colorize.mode(:bold)} to #{pathToFile.colorize.mode(:bold)}"
+      File.write(pathToFile, lineToAdd, mode: "a")
+    end
   end
 end
